vacation-planner-v2/requirements.txt
google-adk>=1.10.0
mcp-flight-search
python-dotenv
httpx
google-adk[eval]